+++ test/conftest.py
@@ -1,14 +1,16 @@
 """Configure pytest."""
 import asyncio
+import os
 import platform
 import sys
 from collections import deque
 from threading import enumerate as thread_enumerate
+from unittest import mock
 
 import pytest
 
 from pymodbus.datastore import ModbusBaseSlaveContext
-from pymodbus.transport import NULLMODEM_HOST, CommParams, CommType
+from pymodbus.transport import NULLMODEM_HOST, CommParams, CommType, ModbusProtocol
 from pymodbus.transport.transport import NullModem
 
 
@@ -93,6 +95,40 @@ def prepare_commparams_client(use_port, use_host, use_comm_type):
         parity="E",
         stopbits=2,
     )
+
+
+
+
+@pytest.fixture(name="client")
+def prepare_protocol(use_clc):
+    """Prepare transport object."""
+    transport = ModbusProtocol(use_clc, False)
+    transport.callback_connected = mock.Mock()
+    transport.callback_disconnected = mock.Mock()
+    transport.callback_data = mock.Mock(return_value=0)
+    if use_clc.comm_type == CommType.TLS:
+        cwd = os.path.dirname(__file__) + "/../examples/certificates/pymodbus."
+        transport.comm_params.sslctx = use_clc.generate_ssl(
+            False, certfile=cwd + "crt", keyfile=cwd + "key"
+        )
+    if use_clc.comm_type == CommType.SERIAL:
+        transport.comm_params.host = f"socket://localhost:{transport.comm_params.port}"
+    return transport
+
+
+@pytest.fixture(name="server")
+def prepare_transport_server(use_cls):
+    """Prepare transport object."""
+    transport = ModbusProtocol(use_cls, True)
+    transport.callback_connected = mock.Mock()
+    transport.callback_disconnected = mock.Mock()
+    transport.callback_data = mock.Mock(return_value=0)
+    if use_cls.comm_type == CommType.TLS:
+        cwd = os.path.dirname(__file__) + "/../examples/certificates/pymodbus."
+        transport.comm_params.sslctx = use_cls.generate_ssl(
+            True, certfile=cwd + "crt", keyfile=cwd + "key"
+        )
+    return transport
 
 
 @pytest.fixture(name="system_health_check", autouse=True)
