+++ test/sub_transport/conftest.py
@@ -1,13 +1,11 @@
 """Fixtures for transport tests."""
 import asyncio
-import os
 from unittest import mock
 
 import pytest
 
 from pymodbus.transport import (
     CommParams,
-    CommType,
     ModbusProtocol,
 )
 
@@ -46,35 +44,3 @@ class DummyProtocol(ModbusProtocol):
 def prepare_dummy_protocol():
     """Return transport object."""
     return DummyProtocol
-
-
-@pytest.fixture(name="client")
-def prepare_protocol(use_clc):
-    """Prepare transport object."""
-    transport = ModbusProtocol(use_clc, False)
-    transport.callback_connected = mock.Mock()
-    transport.callback_disconnected = mock.Mock()
-    transport.callback_data = mock.Mock(return_value=0)
-    if use_clc.comm_type == CommType.TLS:
-        cwd = os.path.dirname(__file__) + "/../../examples/certificates/pymodbus."
-        transport.comm_params.sslctx = use_clc.generate_ssl(
-            False, certfile=cwd + "crt", keyfile=cwd + "key"
-        )
-    if use_clc.comm_type == CommType.SERIAL:
-        transport.comm_params.host = f"socket://localhost:{transport.comm_params.port}"
-    return transport
-
-
-@pytest.fixture(name="server")
-def prepare_transport_server(use_cls):
-    """Prepare transport object."""
-    transport = ModbusProtocol(use_cls, True)
-    transport.callback_connected = mock.Mock()
-    transport.callback_disconnected = mock.Mock()
-    transport.callback_data = mock.Mock(return_value=0)
-    if use_cls.comm_type == CommType.TLS:
-        cwd = os.path.dirname(__file__) + "/../../examples/certificates/pymodbus."
-        transport.comm_params.sslctx = use_cls.generate_ssl(
-            True, certfile=cwd + "crt", keyfile=cwd + "key"
-        )
-    return transport
